refactor: modularize chargily.go into dedicated model files for enhanced scalability and clarity
- Refactored chargily.go by dividing it into separate files: params.go, responses.go, and types.go.
- Moved these model files into a newly created models package to improve separation of concerns.
- Enhanced maintainability and clarity by organizing related structures and defintions.
